+++ packages/stepper/src/Stepper.tsx
@@ -1,11 +1,6 @@
-import { useState, useEffect, useCallback, PropsWithChildren } from 'react';
-import { useStyles } from './style';
-import { clsx } from '@equinor/fusion-react-styles';
-import { Button, Icon } from '@equinor/eds-core-react';
-import { arrow_back, arrow_forward } from '@equinor/eds-icons';
-import StepPane from './StepPane';
-import StepContent from './StepContent';
+import { useState, useEffect, useCallback, PropsWithChildren, createContext, useContext } from 'react';
 import { findNextAvailable, findPrevAvailable, getSteps } from './utils';
+import StepperContent from './StepperContent';
 
 /** Define the props interface for Stepper component */
 export type StepperProps = {
@@ -15,6 +10,7 @@ export type StepperProps = {
   readonly hideNavButtons?: boolean;
   readonly verticalSteps?: boolean;
   readonly horizontalTitle?: boolean;
+  readonly divider?: boolean;
 };
 
 /** Define the type for StepKey */
@@ -25,8 +21,28 @@ export type StepKey = {
   done: boolean;
 };
 
-/** Define the type for StepDirection */
-type StepDirection = 'next' | 'prev';
+/** Create context for Stepper */
+type StepperContextType = {
+  verticalSteps?: boolean;
+  horizontalTitle?: boolean;
+  canPrev: boolean;
+  canNext: boolean;
+  forceOrder: boolean;
+  stepKeys: StepKey[];
+  currentStepKey: string;
+  activeStepPosition: number;
+  handleChange: (stepKey: string, allSteps: StepKey[]) => void;
+};
+
+const StepperContext = createContext<StepperContextType | undefined>(undefined);
+
+export const useStepperContext = () => {
+  const context = useContext(StepperContext);
+  if (!context) {
+    throw new Error('useStepperContext must be used within a StepperProvider');
+  }
+  return context;
+};
 
 export const Stepper = ({
   children,
@@ -36,8 +52,8 @@ export const Stepper = ({
   hideNavButtons,
   verticalSteps,
   horizontalTitle,
+  divider,
 }: PropsWithChildren<StepperProps>): JSX.Element => {
-  const styles = useStyles();
   /** State to manage step keys, current step key, and active step position */
   const [stepKeys, setStepKeys] = useState<StepKey[]>([]);
   const [currentStepKey, setCurrentStepKey] = useState<string>(activeStepKey);
@@ -46,17 +62,6 @@ export const Stepper = ({
   /** State to manage navigation button availability */
   const [canNext, setCanNext] = useState(true);
   const [canPrev, setCanPrev] = useState(false);
-
-  const stepperContainerClasses = clsx(
-    styles.root,
-    styles.stepperContainer,
-    verticalSteps && styles.verticalStepperContainer,
-  );
-  const stepperClasses = clsx(
-    styles.stepper,
-    verticalSteps && styles.verticalStepper,
-    !verticalSteps && horizontalTitle && styles.horizontalTitleStepper,
-  );
 
   /** Effect to update stepKeys when children change */
   useEffect(() => {
@@ -83,22 +88,6 @@ export const Stepper = ({
     setCanPrev(checkPrevious);
   }, [stepKeys, currentStepKey]);
 
-  /** Callback to find the next or previous step key */
-  const findStepKey = useCallback(
-    (direction: StepDirection) => {
-      const current = stepKeys.find((sk) => sk.key === currentStepKey);
-      if (!current) return;
-
-      const nextNewPosition = findNextAvailable(current.position, stepKeys).step?.position;
-      const nextPrevPosition = findPrevAvailable(current.position, stepKeys).step?.position;
-
-      const newPosition = direction === 'next' ? nextNewPosition : nextPrevPosition;
-      const prevOrNext = stepKeys.find((sk) => sk.position === newPosition);
-      return prevOrNext;
-    },
-    [currentStepKey, stepKeys],
-  );
-
   /** Callback to handle step change */
   const handleChange = useCallback(
     (stepKey: string, allSteps: StepKey[]) => {
@@ -108,62 +97,24 @@ export const Stepper = ({
     [onChange],
   );
 
-  /** Callbacks to handle button clicks in navigation */
-  const handleClickPrev = useCallback(() => {
-    const prevKey = findStepKey('prev');
-    if (!prevKey) return;
-    handleChange(prevKey.key, getSteps(children));
-  }, [children, findStepKey, handleChange]);
-
-  const handleClickNext = useCallback(() => {
-    const nextKey = findStepKey('next');
-    if (!nextKey) return;
-    handleChange(nextKey.key, getSteps(children));
-  }, [children, findStepKey, handleChange]);
-
   return (
-    <div className={stepperContainerClasses}>
-      <div className={stepperClasses}>
-        {!hideNavButtons && (
-          <div className={styles.navigationArrows}>
-            <div className={styles.navigationArrow}>
-              <Button
-                color="primary"
-                variant="ghost_icon"
-                onClick={handleClickPrev}
-                disabled={!canPrev}
-                aria-label="Stepper navigation button - previous step"
-              >
-                <Icon data={arrow_back} />
-              </Button>
-            </div>
-            <div className={styles.navigationArrow}>
-              <Button
-                color="primary"
-                variant="ghost_icon"
-                onClick={handleClickNext}
-                disabled={!canNext}
-                aria-label="Stepper navigation button - next step"
-              >
-                <Icon data={arrow_forward} />
-              </Button>
-            </div>
-          </div>
-        )}
-        <StepPane
-          forceOrder={forceOrder || false}
-          activeStepKey={currentStepKey}
-          activeStepPosition={activeStepPosition}
-          onChange={handleChange}
-          verticalSteps={verticalSteps}
-        >
-          {children}
-        </StepPane>
-      </div>
-      <div className={styles.stepContent}>
-        <StepContent activeStepKey={currentStepKey}>{children}</StepContent>
-      </div>
-    </div>
+    <StepperContext.Provider
+      value={{
+        verticalSteps,
+        horizontalTitle,
+        canPrev,
+        canNext,
+        forceOrder: forceOrder || false,
+        stepKeys: getSteps(children),
+        currentStepKey,
+        activeStepPosition,
+        handleChange,
+      }}
+    >
+      <StepperContent hideNavButtons={hideNavButtons} divider={divider}>
+        {children}
+      </StepperContent>
+    </StepperContext.Provider>
   );
 };